chore: move around files in order to mock
