+++ src/lib/processors/mutate-processor.ts
@@ -2,8 +2,7 @@
 // SPDX-FileCopyrightText: 2023-Present The Pepr Authors
 
 import jsonPatch from "fast-json-patch";
-import { kind, KubernetesObject } from "kubernetes-fluent-client";
-import { clone } from "ramda";
+import { KubernetesObject } from "kubernetes-fluent-client";
 import { MeasureWebhookTimeout } from "../telemetry/webhookTimeouts";
 import { Capability } from "../core/capability";
 import { shouldSkipRequest } from "../filter/filter";
@@ -12,11 +11,13 @@ import { AdmissionRequest, Binding } from "../types";
 import Log from "../telemetry/logger";
 import { ModuleConfig } from "../types";
 import { PeprMutateRequest } from "../mutate-request";
-import { base64Encode, convertFromBase64Map, convertToBase64Map } from "../utils";
+import { base64Encode } from "../utils";
 import { OnError } from "../../cli/init/enums";
 import { resolveIgnoreNamespaces } from "../assets/webhooks";
 import { Operation } from "fast-json-patch";
 import { WebhookType } from "../enums";
+import { decodeData, reencodeData } from "../decode-utils";
+
 export interface Bindable {
   req: AdmissionRequest;
   config: ModuleConfig;
@@ -60,32 +61,32 @@ export function logMutateErrorMessage(e: Error): string {
   }
 }
 
-export function decodeData(wrapped: PeprMutateRequest<KubernetesObject>): {
-  skipped: string[];
-  wrapped: PeprMutateRequest<KubernetesObject>;
-} {
-  let skipped: string[] = [];
+// export function decodeData(wrapped: PeprMutateRequest<KubernetesObject>): {
+//   skipped: string[];
+//   wrapped: PeprMutateRequest<KubernetesObject>;
+// } {
+//   let skipped: string[] = [];
 
-  const isSecret = wrapped.Request.kind.version === "v1" && wrapped.Request.kind.kind === "Secret";
-  if (isSecret) {
-    // convertFromBase64Map modifies it's arg rather than returing a mod'ed copy (ye olde side-effect special, blerg)
-    skipped = convertFromBase64Map(wrapped.Raw as unknown as kind.Secret);
-  }
+//   const isSecret = wrapped.Request.kind.version === "v1" && wrapped.Request.kind.kind === "Secret";
+//   if (isSecret) {
+//     // convertFromBase64Map modifies it's arg rather than returing a mod'ed copy (ye olde side-effect special, blerg)
+//     skipped = convertFromBase64Map(wrapped.Raw as unknown as kind.Secret);
+//   }
 
-  return { skipped, wrapped };
-}
+//   return { skipped, wrapped };
+// }
 
-export function reencodeData(wrapped: PeprMutateRequest<KubernetesObject>, skipped: string[]): KubernetesObject {
-  const transformed = clone(wrapped.Raw);
+// export function reencodeData(wrapped: PeprMutateRequest<KubernetesObject>, skipped: string[]): KubernetesObject {
+//   const transformed = clone(wrapped.Raw);
 
-  const isSecret = wrapped.Request.kind.version === "v1" && wrapped.Request.kind.kind === "Secret";
-  if (isSecret) {
-    // convertToBase64Map modifies it's arg rather than returing a mod'ed copy (ye olde side-effect special, blerg)
-    convertToBase64Map(transformed as unknown as kind.Secret, skipped);
-  }
+//   const isSecret = wrapped.Request.kind.version === "v1" && wrapped.Request.kind.kind === "Secret";
+//   if (isSecret) {
+//     // convertToBase64Map modifies it's arg rather than returing a mod'ed copy (ye olde side-effect special, blerg)
+//     convertToBase64Map(transformed as unknown as kind.Secret, skipped);
+//   }
 
-  return transformed;
-}
+//   return transformed;
+// }
 
 export async function processRequest(
   bindable: Bindable,
@@ -196,6 +197,7 @@ export async function mutateProcessor(
   // If no capability matched the request, exit early
   if (bindables.length === 0) {
     Log.info(reqMetadata, `No matching actions found`);
+    webhookTimer.stop();
     return response;
   }
